requirements_py310.txt
coloredlogs==15.0.1
cupy==13.6.0
fastrlock==0.8.3
ffmpeg-python==0.2.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
humanfriendly==10.0
imageio==2.37.0
Jinja2==3.1.6
MarkupSafe==2.1.5
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
onnx==1.19.1
onnx-ir==0.1.11
onnxruntime-gpu==1.23.0
onnxscript==0.5.4
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pyreadline3==3.5.4
sympy==1.14.0
torch==2.9.0+cu126
torchaudio==2.9.0+cu126
torchvision==0.24.0+cu126
typing_extensions==4.15.0
